// plugwerk-server/plugwerk-server-frontend/src/api/generated/model/plugin-paged-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Plugwerk API
 * Plugin marketplace for the Java/PF4J ecosystem
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PluginDto } from './plugin-dto';

/**
 * 
 * @export
 * @interface PluginPagedResponse
 */
export interface PluginPagedResponse {
    /**
     * 
     * @type {Array<PluginDto>}
     * @memberof PluginPagedResponse
     */
    'content': Array<PluginDto>;
    /**
     * 
     * @type {number}
     * @memberof PluginPagedResponse
     */
    'totalElements': number;
    /**
     * 
     * @type {number}
     * @memberof PluginPagedResponse
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof PluginPagedResponse
     */
    'size': number;
    /**
     * 
     * @type {number}
     * @memberof PluginPagedResponse
     */
    'totalPages': number;
}